10/08

+++ VipServices2020.Domain/Models/Price.cs
@@ -8,26 +8,26 @@ namespace VipServices2020.Domain.Models {
         public int Id { get; set; }
         public int FirstHourPrice { get; set; } = 0;
         public int SecondHourCount { get; set; } = 0;
-        public decimal SecondHourPrice { get; set; } = 0;
+        public double SecondHourPrice { get; set; } = 0;
         [NotMapped]
         public readonly int SecondHourPercentage = 65;
         public int NightHourCount { get; set; } = 0;
-        public decimal NightHourPrice { get; set; } = 0;
+        public double NightHourPrice { get; set; } = 0;
         [NotMapped]
         public readonly int NightHourPercentage = 140;
         public int OvertimeCount { get; set; } = 0;
-        public decimal OvertimePrice { get; set; } = 0;
+        public double OvertimePrice { get; set; } = 0;
         public int FixedPrice { get; set; } = 0;
         [Required]
-        public decimal SubTotal { get; set; }
+        public double SubTotal { get; set; }
         public Staffel Staffel { get; set; }
         [Required]
-        public decimal ExclusiveBtw { get; set; }
+        public double ExclusiveBtw { get; set; }
         [NotMapped]
         public readonly int Btw = 6;
         [Required]
-        public decimal BtwPrice { get; set; }
+        public double BtwPrice { get; set; }
         [Required]
-        public decimal Total { get; set; }
+        public double Total { get; set; }
     }
 }
